+++ application/api/service.py
@@ -206,48 +206,3 @@ def get_executed_result(current_nlq_chain: NLQChain) -> str:
     sql_query_result = current_nlq_chain.get_executed_result_df(all_profiles[current_nlq_chain.profile])
     final_sql_query_result = sql_query_result.to_markdown()
     return final_sql_query_result
-
-
-def get_retrieve_opensearch(env_vars, query, search_type, selected_profile, top_k, score_threshold=0.7):
-    demo_profile_suffix = '(demo)'
-    origin_selected_profile = selected_profile
-    selected_profile = "shopping_guide"
-
-    if search_type == "query":
-        index_name = env_vars['data_sources'][selected_profile]['opensearch']['index_name']
-    else:
-        index_name = env_vars['data_sources'][selected_profile]['opensearch']['index_name'] + "_ner"
-
-    embedding_endpoint = os.getenv('SAGEMAKER_ENDPOINT_EMBEDDING','')
-    if embedding_endpoint:
-        records_with_embedding = create_vector_embedding_with_sagemaker(
-            embedding_endpoint,
-            query,
-            index_name=env_vars['data_sources'][selected_profile]['opensearch']['index_name'])
-    else:
-        records_with_embedding = create_vector_embedding_with_bedrock(
-            query,
-            index_name=env_vars['data_sources'][selected_profile]['opensearch']['index_name'])
-    retrieve_result = retrieve_results_from_opensearch(
-        index_name=index_name,
-        region_name=env_vars['data_sources'][selected_profile]['opensearch']['region_name'],
-        domain=env_vars['data_sources'][selected_profile]['opensearch']['domain'],
-        opensearch_user=env_vars['data_sources'][selected_profile]['opensearch'][
-            'opensearch_user'],
-        opensearch_password=env_vars['data_sources'][selected_profile]['opensearch'][
-            'opensearch_password'],
-        host=env_vars['data_sources'][selected_profile]['opensearch'][
-            'opensearch_host'],
-        port=env_vars['data_sources'][selected_profile]['opensearch'][
-            'opensearch_port'],
-        query_embedding=records_with_embedding['vector_field'],
-        top_k=top_k,
-        profile_name=origin_selected_profile.replace(demo_profile_suffix, ''))
-
-    selected_profile = origin_selected_profile
-
-    filter_retrieve_result = []
-    for item in retrieve_result:
-        if item["Score"] > score_threshold:
-            filter_retrieve_result.append(item)
-    return filter_retrieve_result
