+++ application/utils/opensearch.py
@@ -2,7 +2,7 @@ import boto3
 from opensearchpy import OpenSearch, RequestsHttpConnection
 from opensearchpy.helpers import bulk
 import logging
-from utils.llm import create_vector_embedding_with_bedrock, retrieve_results_from_opensearch
+from utils.llm import create_vector_embedding_with_bedrock
 
 logger = logging.getLogger(__name__)
 
@@ -120,3 +120,83 @@ def get_retrieve_opensearch(env_vars, query, search_type, selected_profile, top_
             filter_retrieve_result.append(item)
     return filter_retrieve_result
 
+
+def retrieve_results_from_opensearch(index_name, region_name, domain, opensearch_user, opensearch_password,
+                                     query_embedding, top_k=3, host='', port=443, profile_name=None):
+    auth = (opensearch_user, opensearch_password)
+    if len(host) == 0:
+        host = get_opensearch_endpoint(domain, region_name)
+        port = 443
+
+    # Create the client with SSL/TLS enabled, but hostname verification disabled.
+    opensearch_client = OpenSearch(
+        hosts=[{'host': host, 'port': port}],
+        http_compress=True,  # enables gzip compression for request bodies
+        http_auth=auth,
+        use_ssl=True,
+        verify_certs=False,
+        ssl_assert_hostname=False,
+        ssl_show_warn=False
+    )
+    search_query = {
+        "size": top_k,  # Adjust the size as needed to retrieve more or fewer results
+        "query": {
+            "bool": {
+                "filter": {
+                    "match_phrase": {
+                        "profile": profile_name
+                    }
+                },
+                "must": [
+                    {
+                        "knn": {
+                            "vector_field": {  # Make sure 'vector_field' is the name of your vector field in OpenSearch
+                                "vector": query_embedding,
+                                "k": top_k  # Adjust k as needed to retrieve more or fewer nearest neighbors
+                            }
+                        }
+                    }
+                ]
+            }
+
+        }
+    }
+
+    # Execute the search query
+    response = opensearch_client.search(
+        body=search_query,
+        index=index_name
+    )
+
+    return response['hits']['hits']
+
+
+def upload_results_to_opensearch(region_name, domain, opensearch_user, opensearch_password, index_name, query, sql,
+                                 host='', port=443):
+    auth = (opensearch_user, opensearch_password)
+    if len(host) == 0:
+        host = get_opensearch_endpoint(domain, region_name)
+        port = 443
+    # Create the client with SSL/TLS enabled, but hostname verification disabled.
+    opensearch_client = OpenSearch(
+        hosts=[{'host': host, 'port': port}],
+        http_compress=True,  # enables gzip compression for request bodies
+        http_auth=auth,
+        use_ssl=True,
+        verify_certs=False,
+        ssl_assert_hostname=False,
+        ssl_show_warn=False
+    )
+
+    # Vector embedding using Amazon Bedrock Titan text embedding
+    logger.info(f"Creating embeddings for records")
+    record_with_embedding = create_vector_embedding_with_bedrock(query, index_name)
+
+    record_with_embedding['sql'] = sql
+    success, failed = put_bulk_in_opensearch([record_with_embedding], opensearch_client)
+    if success == 1:
+        logger.info("Finished creating records using Amazon Bedrock Titan text embedding")
+        return True
+    else:
+        logger.error("Failed to create records using Amazon Bedrock Titan text embedding")
+        return False
